Refactor Out Shared Code
Some code is shared between different parts, or may be generally useful to have when solving other Advent of Code challenges
