+++ src/main/kotlin/com/geistindersh/aoc/year2015/Day24.kt
@@ -2,6 +2,7 @@ package com.geistindersh.aoc.year2015
 
 import com.geistindersh.aoc.helper.files.DataFile
 import com.geistindersh.aoc.helper.files.fileToStream
+import com.geistindersh.aoc.helper.iterators.subsetSum
 import com.geistindersh.aoc.helper.report
 
 class Day24(dataFile: DataFile) {
@@ -14,21 +15,6 @@ class Day24(dataFile: DataFile) {
 		override fun compareTo(other: Group): Int {
 			if (other.weights.size != weights.size) return weights.size.compareTo(other.weights.size)
 			return quantumEntanglement().compareTo(other.quantumEntanglement())
-		}
-	}
-
-	private fun List<Int>.subsetSum(target: Int): Sequence<List<Int>> = sequence {
-		if (target == 0) {
-			yield(emptyList())
-			return@sequence
-		}
-		val options = this@subsetSum
-		options.forEachIndexed { index, option ->
-			if (option <= target) {
-				this@subsetSum.subList(index + 1, options.size)
-					.subsetSum(target - option)
-					.forEach { yield(listOf(option) + it) }
-			}
 		}
 	}
